+++ test/persistence/DummyMemoryPersistence.go
@@ -116,7 +116,7 @@ func (c *DummyMemoryPersistence) GetPageByFilter(correlationId string, filter *c
 	return page, err
 }
 
-func (c *DummyMemoryPersistence) GetCountByFilter(correlationId string, filter *cdata.FilterParams) (count int, err error) {
+func (c *DummyMemoryPersistence) GetCountByFilter(correlationId string, filter *cdata.FilterParams) (count int64, err error) {
 
 	if &filter == nil {
 		filter = cdata.NewEmptyFilterParams()
